Moved after submission email sending task to post processor thread.
